HACK: Add hard-coded May 2023 social media and live stream dates
For the May 2023 cohort, we are planning to host two live streams for
mentors, and two live streams for applicants.

I don't have time to figure out a model for these chats. The simple way
would be to add more fields to the RoundPage class. However, it's very
likely we will change chat formats, number of chats, and when they are
held. So for now, I won't make model changes, and I'll just hard-core
them into the applicant and mentor time line snippets.

Note that this will make them appear on the past internship cohort
pages. Oh well? I'll fix this some time in the future. (Sorry future
self!)