Add week 6 email - review internship timeline
Add the careers chat discussion description to the internship guide.

Move the intern and mentor conversation topic about reviewing the
timeline to week 6.

For the December 2022 cohort only: move the careers chat to week 6.
The following cohorts will have that chat in week 7. The chat section of
the week 6 email will be removed after I send it to the mentors and
interns.