Fix date to send reminder about career chat email
We moved the careers chat 2 weeks earlier in the schedule. Therefore we
need to send the reminder out sooner as well.

Also correct the code so that it doesn't email the opportunities and
mentor mailing lists multiple times. The past bug would have emailed the
lists for as many interns we had (i.e. 65 times!).