Add automated email to tell past mentoring orgs about the community CFP
When the Outreachy call for mentoring organizations goes out, we send it
to the mentor mailing list and the announcement mailing list. However,
mentors either ignore it because it's sent close to the beginning of the
year (in January for the May cohort), or don't see it because Gmail
filters emails to mailing lists into the Promotions or Updates tab.

Send a personalized email to mentoring organization coordinators.
Remind them of the community CFP deadlines (mentoring organization
sign-up deadline, and project submission deadline). Exclude any
mentoring organizations that have already signed up to participate in
the current cohort, or indicated that they will not participate in the
current cohort, or have been rejected by the Outreachy organizers to
participate in the current cohort.

Add some hard-coded dates for live-streamed mentor chats. If the chats
work out this cohort, we can add the dates as a RoundPage field. Since
they're an experiment this cohort, don't add the fields just yet.