Add link to open Conservancy Outreachy invoice queue
Add a link to open Conservancy's RT ticket queue for the
accounts-receivable queue, filtered for the Outreachy project, and
sorted by the newest sent invoice.

Make the link open in a new tab or window, using the target="blank_"
html tag attribute.

This will help us when we're trying to figure out whether a sponsorship
request has been sent to Conservancy, or what the status of an invoice
request is.