Show internship contacts for only active internships.
TLDR: database design technical debt is hard. We'll work around bad
database design by changing the view's database query, and switching
which rare corner case is hit.

When this code was originally designed, I assumed mentors would always
give final feedback through the website and Outreachy organizers would
always set the organizer_approved flag to True once an internship final
payment authorization had been sent.

First, it turns out I am terrible at setting the final payment
authorization flag for interns. I send the authorizations, but I'm so
burned out from trying to run an internship cohort and get the
applications open for the next cohort that I forget to set the flag for
each intern. Interns get paid, records get submitted to Conservancy, but
I forget to set the final payment flag in the Outreachy website.

It also turns out there are rare some cases where a mentor thinks the
intern should not be paid the final internship stipend, and the
Outreachy organizers disagree.

This results in the mentor not submitting final feedback to the
Outreachy website. Instead, we have a conversation with the mentor over
email. This often results in an internship ending successfully, or the
intern completing their internship under a different mentor. We end up
submitting documentation to Conservancy manually to approve the intern
payment.

The end result is the intern has been paid, but there is no
FinalMentorFeedback object to set the organizer_payment_approved flag
in. To fake an object just to set a flag would violate what we're trying
to do with the FinalMentorFeedback object -- which is to keep a
revision-controlled record of what the mentor's feedback was.

This means the second query (that I've deleted) will return the intern
and mentor contacts for these rare cases in the active internship
contacts list.

The active internship contacts are used to add new mentors and interns
to the appropriate mailing lists and the Outreachy Zulip chat.

Since in this case the Outreach organizers disagree with the mentor's
judgment on internship payment, the mentor may have decided to
unsubscribe from the Outreachy mentor mailing or deactive their
Outreachy Zulip chat account. We definitely don't want to try to
re-subscribe them or resent them a Zulip chat invite.

The end result is we can't trust the organizer_payment_approved flag in
the FinalMentorFeedback object to accurately represent whether an
internship has been completed.

For now, simplify this view by only showing internships that are active
based on internship dates in the InternSelection object.

This does mean that interns who need an internship extension (where
their individual internship end date hasn't been updated yet) won't show
up in this list. I think that's fine. The rare case we miss is when an
internship hasn't been extended yet, and a new mentor needs to be
subscribed to the mailing list and Zulip chat. But in that case, we can
look them up manually, or change the internship end date.

I need to address the underlying issue through a database change to move
the Outreachy organizer's internship final payment approval out of the
FinalMentorFeedback object. Ideally, it would also move out of the
MidpointMentorFeedback object as well, since that can also be overridden
by Outreachy organizers. That means BaseMentorFeedback needs to change
too, I think?

The longer-term solution will make revision tracking of the mentor
changes to those two feedback object much easier, since we won't need to
filter out Outreachy organizer changes to those objects.