Update Python packages except Django, Wagtail, django-formtools, and django-reversion
I'm going to slowly work my way through the new versions of various
libraries to see how far we can update the website.

django-formtools might have a fix for the regression they introduced in
2.4 for checking conditionals to prune form pages. The potential fix was
introduced in 2.4.1:

https://github.com/jazzband/django-formtools/pull/239

I have to see if their fix works for our usage case, or if we need to
work-around the issue further and/or ditch django-formtools altogether.

This is getting very urgent, as Django 3.2 will be end-of-life in 2024.
We have to implement new initial application code if the bug we ran into
isn't fixed in django-formtools 2.4.1. This would mean re-writing the
initial application form to save the form classes at every step. That
would be a very big re-write, and I do not think we can do it before the
initial application period opens in January 2024.

Fingers crossed updating fixes the issue. I don't have high hopes
though, since the django-formtools folks seem to have only implemented a
partial fix:

https://github.com/jazzband/django-formtools/pull/239#issuecomment-1542931733

[Update: the fix works!]