Update django-reversion Python package
I tested locally to see that the Sponsorship objects under revision
control would still be tracked properly.

I could see that when I made a change in the Django admin interface to
the Sponsorship object that the change was reflected in the revision
history.

I also tried reverting an object back to before the change. That worked
fine, and the object's history showed the change and the revert as two
revision control history points.