Update pytz Python package to latest
We only use the pytz package for generating the Outreachy blog RSS feed.
It's used to output the date and time and timezone of when the blog post
was first written, and last updated.

The pytz Python package has been replaced with native timezone support
in Python 3.9. See the Note in the Introduction paragraph on the pytz
pypi page:

https://pypi.org/project/pytz/

The Outreachy website is already running Python 3.9.17 in production.

Theoretically, we should be able to use the native Python datetime
module to represent this timezone aware datetime. However, the native
Python datetime module lacks a localize() function like pytz has.

I started converting the code to native Python, and I found myself
getting confused as to what type of timezone Wagtail needs (if any).

Things are further confused by the fact that the new native support
needs to know whether you are asking for a timezone that is currently in
daylight savings time or standard time. E.g. PDT (UTC-8) or PST (UTC-7).
I hate daylight savings time. Please please make it go away.

After getting a headache, I threw up my hands and updated the pytz
Python package, rather than trying to replace it with native Python
support.

I downloaded the RSS feed both before and after the pytz package upgrade
to ensure that the ATOM files were identical.