WIP: remove essay rating system from website
We've shifted over to an essay qualities system, and moved away from
rating essays.

I first approached this work by wanting to remove the unused fields,
reorganize the database models, and remove unused URLS.

However, this is the first cohort we're trying this new essay reviewing
methodology, and it's unclear whether this new system will be used in
the future. We may go back to rating essays with more precise
instructions, or guided by our new essay qualities.

So it's unclear whether we will need the essay rating models in the
future, and it's a whole lot of work to rip out all the URLs and view
calls.

I'm going to try this again, with more focus on working with the data
structures we currently have. Then I'll clean up the code in June or
July and remove the ratings system if we need to.