Clarify commands to start local Django webserver
I think newer versions of postgres changed the port number it runs
under. I also think postgres requires that users provide their password
to connect to the database (even when the user owns the database).

Therefore, we need to ask postgres to prompt us for a password during
user creation. We also need to supply the password and new port number
when connecting to postgres.