Move EssayQuality category names into new EssayQualityCategory objects
Add a new field to the EssayQuality class, category, that points to one
EssayQualityCategory object. If the EssayQualityCategory object is
deleted, cascade the delete and remove all EssayQuality objects from
that category.

Create new EssayQualityCategory objects for each category name already
in the database. Standardize the name to be lower case with a
capitalized first letter.

Set the category object pointer in each EssayQuality object to the
correct new EssayQualityCategory object.