Remove section on community "credits" from documentation
Outreachy is an initiative under a non-profit 501c3 organization
(Software Freedom Conservancy). Conservancy leaders are concerned that
the Outreachy program website and documentation are not aligned with
Conservancy accounting procedures and standard non-profit accounting
language. This commit begins the word of aligning the Outreachy website
and documentation with Conservancy's accounting processes.

This commit removes an outdated section of the documentation for open
source communities that mentor Outreachy interns. Section 3.2 said that
open source communities would receive a "credit" if a sponsor was
invoiced for a donation towards an Outreachy internship with a specific
open source community, and the community didn't pick an intern that
cohort. The section also said that the open source community would
receive a partial credit for a future Outreachy internship if a current
internship was terminated without paying the final internship stipend.

However, the word "credit" could imply that a donor is "buying" an
Outreachy internship for an open source community -- which they are not.

Sponsorship of Outreachy is a charitable donation to a US-based
non-profit. Donations to Outreachy are used to support Outreachy interns
working with open source communities, to support Conservancy staff and
contractors who work with mentors, applicants, and interns, and to
promote the Outreachy program and open source to people who face
systemic bias and discrimination in the technology industry of country
they are living in.

Some donors may ask to direct their donation towards Outreachy
internships with specific open source communities. Conservancy and
Outreachy encourage donors to gave to the Outreachy program general
activities fund instead. This provides us more flexibility, and helps
simplify our bookkeeping.

Still, there are cases where donors are not able to give to Outreachy
program activities generally, and are only able to give towards
Outreachy internships with specific open source communities. An example
might be when an open source community directs the non-profit
organization who holds community funds to donate towards Outreachy
interns with that open source community. In this case, the open source
community's fiscal sponsor might insist that its donation only go
towards activities that further their non-profit mission -- namely
Outreachy internships with a specific open source community.

If a donor explicitly instructs us in writing to direct their donation
towards Outreachy internships with a specific open source community,
Conservancy leaders will discuss the donation internally. If the
donation is aligned with Conservancy's non-profit mission and the open
source community meets Outreachy's eligibility criteria, Conservancy
will send an invoice to the donor. Once the invoice is paid, Conservancy
will record that donation as a "temporarily restricted earmark" for a
"donor restricted fund".

Outreachy sponsors should refer to their donation invoice description
for information about the specific Outreachy activities their restricted
donations will be used for. Outreachy sponsors should read the invoice
terms for more details about the amount of time their directed donation
will be temporarily restricted towards those activities.

Conservancy may change the language and terms on an Outreachy sponsor
invoice from time to time. Therefore, that language should be documented
in the invoice, not in the Outreachy website. Otherwise we run the risk
of the documentation going stale.