Correct language around Outreachy sponsorship leads
WIP that needs implementing:
 - create an email that is sent to Outreachy organizers when a
   coordinator updates the sponsorship leads
 - edit the way each field in the sponsorship leads form is displayed,
   adding the field description and help text that was agreed upon by
   Conservancy leadership
 - investigate why not all of the community sign up tests are passing

Outreachy is an initiative under the USA non-profit Software Freedom
Conservancy. This language change has been discussed by Conservancy
leadership, and Sage Sharp has provided this description of those
discussions to the best of their understanding.

USA charitable organization rules state that non-profit staff cannot
direct donations towards a particular purpose. Instead, the donor must
provide any donation instructions themselves. This means, for example,
that if a donor wants to give to the organization generally, the
charitable staff cannot direct the donation towards a more restrictive
temporarily earmarked fund.

As USA charitable organization staff and contractors, Outreachy
organizers must assume any organization wanting to sponsor Outreachy
wants the least restrictive donation.

By default, Outreachy and Conservancy must assume a sponsor wants to
donate to the Outreachy general activities fund. This is the least
restrictive earmarked fund available to Outreachy sponsors. The only
less restrictive donation would be to sponsor Conservancy, rather than
one of Conservancy's initiatives (Outreachy), but that would make the
donor a sponsor of Conservancy, not a sponsor of Outreachy.

Outreachy and Conservancy must also assume that if a sponsor says they
want to donate to "Outreachy internships" then the sponsor must want to
donate towards all Outreachy internships with any community. This is
less restrictive than donating towards Outreachy internships with a
particular FOSS community.

Supporting donations that are directed in a more restrictive way than
towards Outreachy general activities or towards Outreachy internships
with any organization adds complexity to Conservancy's accounting
bookkeeping. As Conservancy is a small organization, both Conservancy
and Outreachy organizers want to reduce overhead on our accounting
staff. Thus we are working together to encourage sponsors to give more
generally, either towards the Outreachy general activities fund or
towards Outreachy internships with any community.

If a sponsor wants to donate towards Outreachy internships with a
particular FOSS community, they need to provide explicit instructions
(preferably in writing).

Additionally, in order to support Conservancy's accounting staff,
Outreachy organizers are attempting to gather more information up front
from community coordinators. This information will help us determine
what extra steps we need to take for each sponsorship, and help us
prioritize which sponsorship leads to act on first.

For example, it is hard for us to determine whether a sponsorship lead
has been given to us by a representative of the sponsoring organization,
or if a person that is not associated with the sponsoring organization
is saying they might know of someone who wants to donate to Outreachy.

Outreachy sponsorship leads are provided by FOSS community coordinators
when they sign up to participate in Outreachy.

In some cases, the community coordinator is a representative of the
sponsoring organization. For example, the Treasurer of Yet Another
Society (The Perl Foundation) might sign up Perl and Raku to mentor
Outreachy interns. In that case, the Treasurer is acting as a fiscal
representative of Yet Another Society, telling Outreachy where to direct
its donation.

In other cases, the community coordinator is not a representative of the
sponsoring organization. For example, a Debian developer might sign
Debian up to mentor Outreachy interns. That Debian developer provides
Outreachy a sponsorship lead that Debian may want to donate some of its
funds that are held by Software in the Public Interest (SPI) towards
Outreachy. However, that Debian developer is not a fiscal representative
of the Debian Project. Only the Debian Project Leader (DPL) can direct
SPI to donate to Outreachy, and only the DPL can provide instructions as
to where the donation should be directed.

This is just an example of some of the complexity that Outreachy
organizers and Conservancy face as we try to improve the Outreachy
sponsorship leads gathering processes. Please have patience with us as
we try to reduce the overhead for Conservancy accounting staff, and act
on urgent sponsorships or sponsorships that require a long set-up time
more quickly.