Sponsorship leads: finalize form, email organizers on submission
Finalize the fields and descriptions for the form community coordinators
use to give us leads about potential Outreachy sponsors. Some fields
needed longer text fields, so I switched them from a CharField to a
TextField. Some date fields are now optional, and should be saved as
blank dates (whereas before they were automatically saved as today's
date by default, which is not useful).

Send an email to Outreachy organizers whenever a coordinator edits the
form. Note that one community coordinator (who is a representative of
the sponsor) may fill out the form initially, and a second community
coordinator (who is not a representative of the sponsor) could then edit
the form. So it's useful to record in the email both who edited the
sponsorship info, and what sponsor relationship was originally provided.

Note that this will email Outreachy organizers if the sponsorship
information is deleted.