Fix test for community sign up sponsorship leads form
When creating the community sign up test, I think I forgot to fill in
some required fields in the Sponsorship form. That caused the
Sponsorship object to not get saved (because the form was throwing an
error that was not corrected).

The form error wasn't being printed by the test harness code, because I
had set follow=True in the post request for the form. That meant the
code followed the redirect on form save to the next form page (Outreachy
general funding form). The end result was no sponsorship object saved,
and a 200 response code *from the Outreachy general funding form* which
has no required fields.

I've fixed the bugs in the test, and it looks like sponsorship
information gets saved correctly.