Removal of the Opportunities page
We haven't updated our Opportunities page in several years. This page is a legacy page, possibly created by Marina Zhurakhinskaya. The numbers in the headings of this page possibly correspond to the amount of interns sponsored by each supporting organization cited in this page. We used to use it to point applicants to other opportunities in free and open source projects and communities, but it hasn't been used for that purpose in years. Taking all of that into account, we've decided to retire this page.